+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- The next version number should be the version already set in pubspec.yaml -->
 ### Fixed
 - WEB: Fix parsing survey rules [#5083](https://github.com/rokwire/illinois-app/issues/5083).
+### Changed
+- WEB: Introduce image cache service and web network image [#5240](https://github.com/rokwire/illinois-app/issues/5240).
 
 ## [1.10.3] - 2025-06-23
 ### Changed